matrix.h, mission.h, technique.h: Fix a few more repetitious warnings

+++ engine/src/cmd/script/mission.h
@@ -322,7 +322,7 @@ public:
     missionNode *defvar_node;
     missionNode *block_node;
 
-    unsigned int varId;
+    size_t varId;
 };
 
 /* *********************************************************** */
